+++ src/GRA.Controllers/HomeController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using EmailValidation;
 using GRA.Controllers.Attributes;
 using GRA.Controllers.ViewModel.Home;
 using GRA.Domain.Model;
@@ -108,14 +109,24 @@ namespace GRA.Controllers
         {
             if (!string.IsNullOrEmpty(viewModel?.Email))
             {
-                var currentCultureName = _userContextProvider.GetCurrentCulture()?.Name;
-                var currentLanguageId = await _languageService
-                    .GetLanguageIdAsync(currentCultureName);
-                await _emailReminderService
-                    .AddEmailReminderAsync(viewModel.Email,
-                    viewModel.SignUpSource,
-                    currentLanguageId);
-                ShowAlertInfo(_sharedLocalizer[Annotations.Info.LetYouKnowWhen], "envelope");
+                var email = viewModel.Email.Trim();
+                if (EmailValidator.Validate(email))
+                {
+                    var siteStage = GetSiteStage();
+                    var currentCultureName = _userContextProvider.GetCurrentCulture()?.Name;
+                    var currentLanguageId = await _languageService
+                        .GetLanguageIdAsync(currentCultureName);
+                    await _emailReminderService
+                        .AddEmailReminderAsync(email,
+                        siteStage.ToString(),
+                        currentLanguageId);
+                    ShowAlertInfo(_sharedLocalizer[Annotations.Info.LetYouKnowWhen], "envelope");
+                }
+                else
+                {
+                    ShowAlertWarning(_sharedLocalizer[Annotations.Validate.EmailAddressInvalid, 
+                        email]);
+                }
             }
             return RedirectToAction(nameof(Index));
         }
@@ -689,7 +700,6 @@ namespace GRA.Controllers
             switch (siteStage)
             {
                 case SiteStage.BeforeRegistration:
-                    viewmodel.SignUpSource = nameof(SiteStage.BeforeRegistration);
                     if (site != null)
                     {
                         viewmodel.CollectEmail = await _siteLookupService
@@ -716,7 +726,6 @@ namespace GRA.Controllers
                     return View(ViewTemplates.ProgramEnded, viewmodel);
 
                 case SiteStage.AccessClosed:
-                    viewmodel.SignUpSource = nameof(SiteStage.AccessClosed);
                     if (site != null)
                     {
                         viewmodel.CollectEmail = await _siteLookupService