+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - View of and access to historical report output
 - QR code generation ability
 - Spanish localization for jQuery validation messages
+- Email address validation
+- Cannot be emailed property to users
 - Participant personal point goal
 - Join codes for tracking signups
 - Ability to override activity button text
@@ -34,6 +36,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - Display of modals after Bootstrap 5.3 upgrade
 - Performer program cost to display in currency format
+- Email reminder signup source being editable in html
 - Missing translations for secret codes, books and events
 - Password change confirmation not functioning
 - Challenge category multiselect formatting for participants